src/requirements.txt
googleapis_common_protos
google-cloud-tasks
google-cloud-ndb
enum34
Flask==1.1.2
grpcio==1.0.0
sendgrid